Feat/better event signature (#58)
* [feature] Event signature TinyMCE

Users can now add some content under the image of the event signature.
This TinyMCE has the ability to make hyperlinks.

* [feature] Image href event signature

The image of the event signature can now redirect to anything after clicking on it.
An input under the one to choose the image link has been added

Also, I fixed a bug when changing from event to basic signature after changing one of the two inputs, when coming back to event signature, the value was the previous one. It's now fixed.